Add Calibration

+++ app/src/main/java/com/example/studioprojektowe2/coordinates/Acceleration.java
@@ -3,53 +3,53 @@ package com.example.studioprojektowe2.coordinates;
 
 public class Acceleration {
 
-    private Float a_x = 0.0F;
-    private Float a_y = 0.0F;
-    private Float a_z = 0.0F;
+    private Double a_x = 0.0;
+    private Double a_y = 0.0;
+    private Double a_z = 0.0;
 
     public Acceleration() {}
 
-    public Acceleration(Float a_x, Float a_y, Float a_z) {
+    public Acceleration(Double a_x, Double a_y, Double a_z) {
         this.a_x = a_x;
         this.a_y = a_y;
         this.a_z = a_z;
     }
 
-    public Float getA_x() {
+    public Double getA_x() {
         return a_x;
     }
 
-    public void setA_x(Float a_x) {
+    public void setA_x(Double a_x) {
         this.a_x = a_x;
     }
 
-    public Float getA_y() {
+    public Double getA_y() {
         return a_y;
     }
 
-    public void setA_y(Float a_y) {
+    public void setA_y(Double a_y) {
         this.a_y = a_y;
     }
 
-    public Float getA_z() {
+    public Double getA_z() {
         return a_z;
     }
 
-    public void setA_z(Float a_z) {
+    public void setA_z(Double a_z) {
         this.a_z = a_z;
     }
 
     public void setAccelerationTo0() {
-        this.a_x = 0.0F;
-        this.a_y = 0.0F;
-        this.a_z = 0.0F;
+        this.a_x = 0.0;
+        this.a_y = 0.0;
+        this.a_z = 0.0;
     }
 
-    public float[] toArray() {
-        return new float[]{a_x, a_y, a_z};
+    public double[] toArray() {
+        return new double[]{a_x, a_y, a_z};
     }
 
-    public void readFromArray(float[] data) {
+    public void readFromArray(double[] data) {
         this.a_x = data[0];
         this.a_y = data[1];
         this.a_z = data[2];
