+++ app/src/main/java/com/example/studioprojektowe2/coordinates/Distance.java
@@ -3,52 +3,52 @@ package com.example.studioprojektowe2.coordinates;
 
 public class Distance {
 
-    private Float x = 0.0F;
-    private Float y = 0.0F;
-    private Float z = 0.0F;
+    private Double x = 0.0;
+    private Double y = 0.0;
+    private Double z = 0.0;
 
     public Distance() {}
 
-    public Distance(Float x, Float y, Float z) {
+    public Distance(Double x, Double y, Double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public Float getX() {
+    public Double getX() {
         return x;
     }
 
-    public void setX(Float x) {
+    public void setX(Double x) {
         this.x = x;
     }
 
-    public Float getY() {
+    public Double getY() {
         return y;
     }
 
-    public void setY(Float y) {
+    public void setY(Double y) {
         this.y = y;
     }
 
-    public Float getZ() {
+    public Double getZ() {
         return z;
     }
 
-    public void setZ(Float z) {
+    public void setZ(Double z) {
         this.z = z;
     }
 
-    public void updateDistance(Acceleration acceleration, Float time, Velocity velocity) {
-        this.x = (velocity.getV_x() * time) + (0.5F * acceleration.getA_x() * time * time);
-        this.y = (velocity.getV_y() * time) + (0.5F * acceleration.getA_y() * time * time);
-        this.z = (velocity.getV_z() * time) + (0.5F * acceleration.getA_z() * time * time);
+    public void updateDistance(Acceleration acceleration, Double time, Velocity velocity) {
+        this.x = (velocity.getV_x() * time) + (0.5 * acceleration.getA_x() * time * time);
+        this.y = (velocity.getV_y() * time) + (0.5 * acceleration.getA_y() * time * time);
+        this.z = (velocity.getV_z() * time) + (0.5 * acceleration.getA_z() * time * time);
     }
 
     public void setDistanceTo0() {
-        this.x = 0.0F;
-        this.y = 0.0F;
-        this.z = 0.0F;
+        this.x = 0.0;
+        this.y = 0.0;
+        this.z = 0.0;
     }
 
 }
